requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
click==7.1.2
django-appconf==1.0.4
django-read-only-admin==0.14.14
django==3.1.6
python-dotenv[cli]==0.15.0
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
